TypeScript comment-adverts, dfp-env and get-advert-by-id
