--- static/src/javascripts/projects/commercial/modules/dfp/get-advert-by-id.js
@@ -1,6 +1,0 @@
-import { dfpEnv } from './dfp-env';
-
-const getAdvertById = (id) =>
-    id in dfpEnv.advertIds ? dfpEnv.adverts[dfpEnv.advertIds[id]] : null;
-
-export { getAdvertById };
