+++ static/src/javascripts/projects/commercial/modules/dfp/get-advert-by-id.ts
@@ -1,0 +1,7 @@
+import { dfpEnv } from './dfp-env';
+import type { Advert } from './Advert';
+
+const getAdvertById = (id: string): Advert | null =>
+	id in dfpEnv.advertIds ? dfpEnv.adverts[dfpEnv.advertIds[id]] : null;
+
+export { getAdvertById };
